Adding filter dialog component
Added filter dialog component for use in filtering on the explore page and the calendar page.
Added filter buttons to calendar page but did not yet link them to the dialog.
Added imports for Angular Material components used in the filter dialog component.
Added condition to explore page that requires meals to not be fully booked.